Split AshTags into AshBlockTags and AshItemTags
