+++ common/src/main/java/net/dodogang/plume/mixin/AbstractBlockStateMixin.java
@@ -1,6 +1,6 @@
 package net.dodogang.plume.mixin;
 
-import net.dodogang.plume.ash.tag.AshTags;
+import net.dodogang.plume.ash.tag.AshBlockTags;
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
@@ -26,7 +26,7 @@ public abstract class AbstractBlockStateMixin {
     @Inject(at = @At("HEAD"), method = "isOf", cancellable = true)
     private void isBookshelf(Block block, CallbackInfoReturnable<Boolean> info) {
         if (block.equals(Blocks.BOOKSHELF)) {
-            info.setReturnValue(this.isIn(AshTags.Blocks.BOOKSHELVES));
+            info.setReturnValue(this.isIn(AshBlockTags.BOOKSHELVES));
         }
     }
 }
